Neovim: replace persisted.nvim with auto-session
I came across a situation where having the `persisted.nvim` plugin
loaded resulted in the editor – under somewhat unclear circumstances,
but then those circumstances came around they were persistent – always
closing with a non-zero exit.

I ended up writing about this [on the issue tracker for the plugin][1],
and the author pointed me to the fact that this is most likely related
to a [Neovim issue][2], not a plugin issue.

  [1]: https://github.com/olimorris/persisted.nvim/issues/84
  [2]: https://github.com/neovim/neovim/issues/21856

So I don't really have any strong reason to switch, apart from the fact
that with `persisted.nvim` I'll have to add a workaround for the Neovim
issue (at least, so it seems). Maybe this workaround will end up being
required for `auto-session` as well, not sure.

I guess the upside to switching is that `auto-session` seems to be the
more used plugin, if one is to go by amount of stars.

Anyway, I've made this patch more or less revertable, in case I would
feel like going back to `persisted.nvim` in the future.