Neovim: move in vim-polyglot configuration into lazy.nvim spec
The `before_load_plugin_settings` file is probably still around because
I previously didn't know how to set the relevant plugin values on plugin
startup.

Since I'm only doing this with the `vim-polyglot` plugin, move the
values into the `init` function of the `lazy.nvim` specification for the
`vim-polyglot` plugin.