Neovim: call zen-mode through a lambda instead of a command
I'm not sure why, but since the plugin refactoring (1f5bf2c1), for some
reason the mapping for zen-mode does not actually trigger the loading of
the zen-mode plugin, meaning that the command `:ZenMode` does not get
added in time for the mapping to work.

For some reason, however, using a lambda that calls `toggle` on the
required module works. I'm not sure why this works when the current
setup does not, but I just want to have it fixed, so I'll run with this
solution.