Neovim: make cmp-buffer use iskeyword for word regex
Currently the `nvim-cmp` source `cmp-buffer` (completing words from the
current buffer) is not completing words with Swedish characters, since
the words can have letters which are not a-z.

Example: completion for "belåten" becomes "bel", because the word
"breaks" on the "å". Same goes for many other characters than å/ä/ö.

I'm not sure what I'll miss from the default `keyword_pattern`, but the
setting from the README that does "just use the `iskeyword` value for
words" seems to work good enough.