is_reboot_necessary: change order of output
Change the order so that the version of the currently running kernel
comes before the kernel that has been installed. This makes more sense
to me.

Additionally, change the order of the lines in the code (and the s1/s2
variable names) to be in line with this ordering. This makes no actual
change to the logic, but makes the code easier to understand.