git: add `assets/vendor` to the gitignore_global
Phoenix projects dump a whole lot of stuff into assets/vendor, and these
things pop up in Neovim's Telescope searches, which clutters that view
a lot. I'm guessing it's not a great idea to ignore this directory, even
after these assets have been generated and added to the repository, but
I'll use this ignore for now, and then see if I need to tweak this.