Fish: add basic configuration
Found a listing of nice plugins, installed with Fisher. Not sure if
`fish_plugins` should be included like this, or if some other files
should be included (like the actual plugin source files), but this will
at least let me know which plugins I should install on the other hosts.