Neovim: configure nvim-treesitter-context
The configuration added in this patch is a copy paste of the default
configuration, with some fields removed that I can't imagine myself
changing.

The only change from default configuration is the `multiline_threshold`
value, which this patch sets to 0. While I'm sure that there are
contexts which can span more than one line, in Elixir (what I write the
most of), allowing `multiline_threshold` to be more than 1 comes with
some irritation:

The usual context defining lines are fine: `defmodule`, `def`, `defp`,
etc. However, it seems like Treesitter considers certain lines, e.g.
`alias` and comments, that can follow `defmodule`/`def`/etc., lead to
all of the lines being pulled in as the context. This can lead to
bizarre situations like when a comment that stretches a multiple lines
is visible as long it's off-screen and one is in the greater context
which is the defined by the line above the comment.

`multiline_threshold` set to 1 seems to be working fine for now, so set
it to that for the time being.