Git: add ending newline to fuller-mod pretty
Sometimes the log messages (when styled with `fuller-mod`) don't have
a separating newline between the commit blocks. It seems like this
mainly happens when some of the messages have a signature block at the
end of the commit body.

I'm not sure how to fix this in a prettier way, so I'll just generally
add an extra newline after all of the bodies. One extra newline isn't
that bad, the worst I've seen is 3 empty lines before the next commit.