Neovim: don't load copilot.lua on the first Insert
This came from a copy-paste of the configuration for packer, converted
to lazy.nvim configuration. Maybe packer works differently, where one
needs to run the `cmd` before packer cares about `event`, but with
lazy.nvim, going into Insert mode is sufficient for loading Copilot,
which is not what I want.