Fish: set default theme for `bat`
Went through all of the default themes and this one seemed to be one of
the top ones. Not sure what the difference is between the various
Monokai versions, just picked the one that wasn't extra bright.

This default will not be appropriate in the cases when I have a light
terminal background, so some alias might be necessary for handling that
case better, but that's for a separate patch.