Fish: translate `fgcpm` function from Zsh
With the help of ChatGPT I managed to translate the `fgcpm` function
(from `.zsh_programming_alias`) to a Fish function.

The main difference compared to the Zsh function is the addition of
a feature that I've been wanting for a long time: filtering the commits
to eliminate duplicate commit titles. I wanted to eliminate on duplicate
commit bodies, but this turned out to be a bit too difficult to achieve
(removing the newlines in the bodies, but not the newlines between the
commits), but just filtering on the titles is fine.