Fish: update bindings for fzf plugin
I thought that I could use Control+Shift bindings for the various fzf
plugin commands, but it turns out that this is not an option.

As the Fish documentation says:

    "For example, Alt+W can be written as \ew, and Control+X (^X) can be
    written as \cx. Note that Alt-based key bindings are case sensitive
    and Control-based key bindings are not. This is a constraint of
    text-based terminals, not fish."

Source: https://fishshell.com/docs/current/cmds/bind.html

This lead me to move the mappings to be lowercase and using Alt instead.