Neovim: fix annoying `<space>` behavior from `cutlass.nvim`
For reasons that I don't quite understand at the moment, `cutlass.nvim`
wants to add a black hole register mapping for space in Select-mode,
which overlaps with Visual-mode. I get that the plugin wants to put
black hole register mappings for a lot of keys, but `<space>` in Select
or Visual is beyond my understanding.

Because I'm using `<space>` as my Leader key, this means that when I'm
selecting something with Visual, pressing space will make the cursor
move one character to the right, selecting more, when I just want to
access my `<leader>` mappings.

Excluding this mapping for `cutlass.nvim`. I don't see how this would be
negative for my configuration.