Neovim: add fix for vim-fugitive not working due to missing netrw
Set up a `:Browse` command, to fix problem with `:GBrowse` complaining
about missing `netrw`. `netrw` is disabled by the `nvim-tree` plugin,
and for some reason `vim-fugitive` requires `netrw` to be loaded. A way
around this problem is to set a `:Browse` command.

The command needs to be adjusted, depending on the OS (Linux vs. OSX),
where Linux needs `xdg-open`, while OSX needs `open`. This should
preferably be expressed as host-specific files, but for now this will
do.