Neovim: add melange colorscheme
Recently I tried out the `zenburn` colorscheme, and while I've
historically much preferred the high contrast themes with pastel colors,
this time I found the lower contrast of `zenburn` to be more interesting
to me than it ever has been (not sure why). With that said, `zenburn`
has the downside of not coming with colors for both light and dark
background.

After looking around for a bit for a lower contrast theme, I found
`melange-nvim` to be the one I liked the most.

  ## Miscellaneous notes

1. `melange-nvim` recommends setting `termguicolors` to `true`. This has
   the effect side effect that the list of colors for `vim-mark` that
   I've configured does not work, as all of those colors are only using
   `ctermfg` and `ctermbg`, and not `guifg` and `guibg`. This made me
   realize that just switching from a custom palette to the `extended`
   palette is sufficient to get acceptable colors that work with
   `termguicolors`. I don't remember what I had against the
   default/extended/maximum palettes, and since `extended` seems to look
   OK, this is an easy solution.

2. Add a Telescope mapping for switching between colorschemes, since
   I might want to go back and forth between `PaperColor` and `melange`.

3. Set the colorscheme by invoking the command in `init.lua`, instead of
   in the `config` function for `papercolor-theme`.