Neovim: add dial.nvim cycling of True/False
While this could be something I only add for the file type of Gleam (the
only language I currently program that uses capital letter True/False),
I think that it's OK to add this as a general thing to cycle through.

If I, at some point, add language specific configurations for dial.nvim,
this will be one of the first things to get moved.