Neovim: disable lazy.nvim plugin configuration change detection
By default, Lazy will detect and inform all open Neovim instances when
a configuration change has been done to the Lazy configuration, i.e. the
Lua files (or relevant parts of `init.lua`) that describe each plugin
that should be loaded by Lazy (and with what configuration).

Since every Neovim instance gets notified of changes, this means that
I might be changing some of the dotfiles in one instance of Neovim, and
a completely unrelated instance will now get warnings about changes. Not
only that, the warnings stack, meaning that for every save of the
dotfiles, a new warning will be queued, in _every_ other Neovim
instance.

The way I've understood it, the main point of the feature is that the
`:Lazy` interface will be up-to-date, but I'm almost certain that the
plugins still need a full Neovim restart for the changes to take effect,
meaning that I'll probably be aware that the other instances need to be
restarted.

It might be that something is actually reloaded, and that feature is not
just a nuisance, but for now I'll just disable it.