Neovim: Revert update of obsidian.nvim configuration, and pin plugin
This patch includes a revert of the changes in commit
60dde3d4a344772d94034589e832297670fa543e.

The changes introduced in `v3.7.0` pertaining to searching/completion
for anchors/headers/blocks were introduced in such a way that it breaks
the workflow that resembled the one found in Obsidian, i.e. completing
with either note names or aliases for them. Now, aliases are somewhat
broken (they're there, but they're not inserted as link text properly)
and the anchor data slows completion down and makes the results more
messy.

Discussion with the plugin author is currently ongoing in a GitHub
[issue][1] and its related PRs:

  1: https://github.com/epwalsh/obsidian.nvim/issues/589

In 60dde3d4, the configuration for `obsidian.nvim` was updated to comply
with the format that can be seen in `v3.7.x`, where it was previously
written for `v3.6.x`.

This commit should be reverted once the issues with the latest plugin
versions have been addressed.