Neovim: unpin obsidian.nvim plugin
Up to and including `v3.6.1` of `obsidian.nvim`, the completion of notes and note aliases
for my Obsidian vault was working fine. `v3.7.0` introduced some changes to the
completion algorithm which aimed to add support for anchors and blocks.
This, however, was done in a way that resulted in these extra
completions being added to the completion of notes and aliases, changing
how the completion worked. This made it unusable for me, so I pinned the
version to `v3.6.1`, until this issue would get resolved.

The closing of issue #589 with PR #590 (commit fda8df4) seems to have
made the completion interface better. Not as good, but now it works
acceptably.