Fish: add loading of mise
I use `mise`, but I seem to have forgotten to add the loading of the
tool into the shell. I'm not sure if this is strictly necessary, as the
tool seems to work without it, but the documentation does suggest that
one adds it to the startup configuration of Fish.