Neovim: change Telescope mapping for dropping action
The combination of Shift+Return does not seem to work on OSX with Kitty.
This means that whatever action is bound for Telescope to happen with
Return is the only action that I can use (presuming I only use these two
mappings).

I recall setting these mappings and having no problem with them; they
were most likely set on primo, i.e. on Linux. At the moment, it's
unclear to me whether the problem is with OSX, OSX+Kitty, or something
else.

Anyway, for now, reverse the mappings of Insert-mode, as I've come to
prefer `select_default` to `select_drop` as a default. Also, keep the
Normal-mode mappings to have `select_drop` on Return, so that I still
have some easy way of doing `select_drop`.