dunst, i3: configure dunst to be invoked from i3 with dunstctl
Having the bindings for dunst be located in the `dunstrc` has two
problems.

Firstly, it's a set of mappings that are not located with the rest of
the desktop mappings, or at least where a lot of them are.

Secondly, it seems like some combinations (e.g. Super+Shift+F10) don't
work when expressed in the `dunstrc`.