feat(i3): add mapping on XF86Calculator for activating xscreensaver
I noticed that while UHK can output a keysym for Sleep (`XF86Sleep`),
there's no keysym for `XF86ScreenSaver`. To be fair, I don't know if
`XF86ScreenSaver` is even part of a standard of some kind: I found it on
some page that listed all of the different XF86 codes that are
understood by XFree86.

Anyway, since I don't need a key for having a calculator pop up when
pressing `XF86Calculator`, I'll use it for locking the desktop.