feat(nvim): move from lspconfig-Tailwind to tailwind-tools.nvim
The plugin `tailwind-tools.nvim` has some features that go beyond just
setting up the Tailwind language server. The plugin recommends letting
it take care of the language server setup, so I'll switch from the
`nvim-lspconfig` code.

The `includeLanguages` setting was added due to the `tailwind-tools`
default `includeLanguages` not setting the `elixir` to work as
`html-eex`. This resulted in the language server not actually being
turned in Elixir files, meaning that while the server was active in
template files (`.heex`), if one would want to use Tailwind classes in
inlined templates (e.g. in a `render` function), then the defaults would
not work.

The other languages set (`elixir`, `heex`, `eruby`) to be used as
`html-eex` are probably not necessary, but I'll put this here to make
sure that the languages are invoked correctly.