feat(Neovim): add "minty" plugin for color picking
I'm not sure if I really need this plugin, but I've definitely had many
occasions when I've needed to tweak a color (for HTML or otherwise) and
I've brought out some simple utility like gcolor2. I'm not sure if this
will work better or worse, but it's maybe worth trying.