feat(Neovim): add nvim-highlight-colors and remove lspkind-nvim
I've always needed to interact with colors every now and then, but this
has recently increased to the point that I would like to get an idea of
what color a code corresponds to more easily (at least compared to
copy-pasting the code and entering it to some external tool). The
`nvim-highlight-colors` plugin seems like a good thing to try for this
purpose.

Since `nvim-highlight-colors` highlights tailwind classes, I don't need
to use `lspkind-nvim` to inline a color preview in the `cmp`
completions. `lspkind-nvim` has some visual appeal (with its symbols),
I didn't get the plugin for this purpose.