feat(Neovim): add mappings for toggling Treesitter highlights
I've found that for some filetypes, e.g. Elixir with embedded HEEX
templates, Treesitter does a better job for code highlighting than the
colorscheme I use. This is not the case for all files, and therefore I
don't want to default to using Treesitter.

Add mappings for quick on/off toggling of Treesitter for highlights.