refactor(fish)!: move from z.fish to zoxide
I've recently gotten annoyed by the fish plugin `z`, as it lacks an
interactive mode, i.e. showing a list of directory options to select
from. I appreciated that `z` was implemented in pure fish script, as
I presumed that this would ensure good shell integration (it has
functioned well), but the lack of interactive selection is right now too
annoying to overlook.

Replace the `z` plugin with the `zoxide` plugin.

Additionally, add the `ji` alias function, to match the `zi` alias that
`zoxide` provides (for interactive selection).