fix(Neovim): uncomment required luasnip import lines for eelixir
I'm not sure at this time whether the local imports are actually
necessary, as I believe all snippet files get provided with the full
import list (found at the top of the file).

With that said, the LSP highlighting indicated that it didn't see the
global variables that that were referenced (e.g. `c` for choice nodes),
so at least to make the LSP get less confused, let's uncomment the
relevant `require` statements.