fix(kitty): disable `sync_to_monitor`, to improve suspend-resume freezes
I've had a problem for the past couple of months where various
applications suffer from freezing for a couple of seconds after
a suspend-resume. Kitty was one of these applications, probably
exhibiting this problem the most, and after some investigation I found
someone else on the kitty issue tracker claiming that disabling
`sync_to_monitor` improved their situation.

This seems to be solving the problematic behaviour in kitty for me as
well, leading me to believe that this problem has something to do with
VSync, but I'll have to investigate the other applications further
separately.