refactor(neovim): drop LSP `on_attach` function
The `on_attach` function has two purposes:

1. Provide the correct buffer identifier for the various LSP mappings.
2. Only add LSP commands if an LSP server is actually connected.

I think that the first point is redundant, as it seems that the LSP
commands work without providing the `buf` identifier.

The second point might have some merit, but it's mostly causing surprise
when I rapidly use a mapping, and nothing at all happens, due to the
mapping not existing (or something else happening that overlaps
partially with the mapping).

It seems better if this entire `on_attach` thing is removed, especially
since it currently does nothing but set up mappings.