refactor(neovim): change LSP configuration to use modern format
To my understanding the format in this patch should be functionally
equivalent with the current way of configuring the servers with
`lspconfig` and `cmp_nvim_lsp`.

From what I've tested with ElixirLS, the `capabilities` seem to be
provided to `nvim-cmp` without the need for the `capabilities`. Not sure
if this applies to all the servers, but I'll change this and revert if
problems arise.