feat(neovim): add "browsher.nvim" plugin for opening code in browser
While `fugitive` provides opening code on Github, it does not allow
(AFAIK) me to choose the command with which it does the opening. This
means that it tries to use `open`, which is appropriate on OSX, but
fails Linux (`xdg-open` is what I need).

Additionally, I've noticed that I barely use the functionality of
`fugitive` anymore, as I use lazygit for most things related to git
these days. This might be the start of moving away from the venerable
old plugin.