feat(neovim): set `linebreak` to be on by default
Previously I was sceptical about having `linebreak` on as a default,
since I worried that breaking lines in code might make it difficult to
know whether whether there was extra whitespace at the end/start of
lines.

Having tried `linebreak` for a while in code, this worry seems
unfounded, and reading code actually becomes easier with the option
turned on. It's extra nice with the recently turned on `breakindent` as
well.