feat(neovim): add "alternative.nvim" plugin
While this plugin has some overlap with "dial.nvim", it also has some
features which "dial.nvim" does not have, such as modification of Elixir
code.

The current version of the plugin (571e945) is slightly broken, and I
should probably be pointing to my fork with a fix, but I'll use the main
repo URL, and hope that my fix is pulled soon.