feat(neovim): change tabstop from 4 to 8
This is a change that does not really affect me much at all, since
everything I code in uses spaces instead of tabs.

With this said, it seems like various sources I've read (the most
relevant for me being the style guide of the Go formatter) that tabs are
to be rendered with the width of 8. I know that the Linux kernel does
the same.

Again, this does not matter much to me, so I'll just change it to be in
line with what's common.