feat(neovim)!: remove `colemak.nvim` plugin, and all related configuration
Since adopting the Glove80 for my workflow, together with a keymap that
uses home-row-mods, shifted keys have gotten somewhat annoying to press.

While this is normally not a problem, when a large portion of the
`colemak.nvim` revolves around pressing Shift to access various features
around the UNEI key/arrow cluster, it quickly becomes tiresome.

Additionally, I've found that I have mappings that I have muscle memory
for when on a keyboard with a Shift key on my left pinky, but when I
need to stop and think what the mapping is when not being able to resort
to muscle memory, it takes an annoying amount of time to find the
mapping. Indeed, I've found myself remembering the default mapping
better than my own Colemak mappings.

The keymap I'm using has an extensive symbol layer which helps with
navigation in Neovim, allowing me to access much of the convenience that
I got from `colemak.nvim`.

Let's therefore try to revert to default mappings, and see if that works
out better.

  ## Miscellaneous changes

- `flash.nvim`
    - Remove the disabling of motion flashes. This should now work
      better, with default mappings.
    - Change mappings to defaults.
- Text objects plugin
    - Remove all mapping customization for Colemak
- Remove `vim-wheel` plugin
    - Firstly, the current mappings of that I have for the `vim-wheel`
      plugin conflict with default Vim mappings.
    - Secondly, I was just now reminded of the scrolling that one gets
      from Normal-mode Shift+Up/Down.
    - Let's therefore simply remove this plugin, as I can probably live
      without it.
- Remove various plugin configuration and mappings that are added for
  compatibility with the `colemak.nvim` plugin.