feat(i3)!: remove mod+f-key adjustment of volume
I no longer use keyboards that do not have access to media keys for
adjusting/muting volume: either the keys are physical, or the keyboard
is configurable and I can have the keys on some alternate keymap layer.

Therefore, these mappings are mostly in the way of potential future
mappings that would want to make use of these positions.