feat(neovim)!: add harpoon plugin, move <leader> mappings to make room
I've seen others use harpoon and it seems like it could be an
interesting addition to the way I work.

It seems like harpoon is would benefit from being available on
low-friction mappings. The left keyboard half has almost no mappings
under <leader>, so let's put the harpoon mappings there.

Some minor movements of mappings needed to be done to make room, but
nothing of great consequence, with respect to frequecy of use.