feat(git): add .git-blame-ignore-revs file to repo
Some commits ago I added the reading of a `.git-blame-ignore-revs` file
by default, for checking for commits that should not be considered for
blame.

I just now ran into the issue that some Git plugins for neovim didn't
work properly if this file was not around. Add an empty file, and then
maybe I'll have reason to add some revisions to it later.