feat(neovim): add mappings for line number management
Since I added the `line-numbers.nvim` plugin for having the ability to
see both relative and absolute line numbers at the same time, I lost the
ability to manage the line numbers with the usual commands.

This means that the unimpaired mappings are of no use, and are therefore
removed in this patch.

Additionally, add mappings under a leader key namespace for enabling
quick management of the line numbers, using the plugin.